Point at our own version of phantomjs2 binary
PhantomJS 2 is not distributing official binary packages for linux right
now. We got this binary from TravisCI - see
https://github.com/travis-ci/travis-ci/issues/3225 - rather than pulling
a binary from their S3 we felt more secure hosting it ourselves. We also
modified the structure of the archive to more closely resemble an
official distribution so that we didn't need to make bigger changes to
the cookbook.

We also removed the x86 section as it pointed at an outdated version on
a defunct google code repository.